Update Version 3.3.5

+++ commands/Moderation/ban.js
@@ -1,5 +1,7 @@
 const Discord = require("discord.js");
 const ms = require('ms');
+const send = require("../../utils/sendMessage.js")
+const getMember = require("../../utils/getMember.js");
 
 module.exports = {
     name: "ban",
@@ -11,54 +13,56 @@ module.exports = {
     category: "Moderation",
     permission: ["BAN_MEMBERS"],
     botreq: "Embed Links, Ban Members, Manage Message",
-    run: async (bot, message, args) => {
+    options: [{
+        name: "user",
+        description: "For which command should I send information for?",
+        required: true,
+        type: 3, //https://discord.com/developers/docs/interactions/application-commands#application-command-object-application-command-option-structure
+        req: "string"
+    },
+    {
+        name: "reason",
+        description: "Reason that why the user is getting banned",
+        required: false,
+        type: 3, //https://discord.com/developers/docs/interactions/application-commands#application-command-object-application-command-option-structure
+        req: "reason"
+    }
+    ],
+    run: async (bot, message, args, options, author) => {
 
         if (!message.guild.me.permissions.has(["BAN_MEMBERS"])) {
             const embed = new Discord.MessageEmbed()
             embed.setColor(0xFF0000)
             embed.setDescription("❌ Check My Permissions. [Missing Permission:- Ban Members]")
-            return message.channel.send({ embeds: [embed] })
+            return send(message, {
+                embeds: [embed],
+                ephemeral: true
+            }, true)
         }
 
         if (!message.guild.me.permissions.has(["MANAGE_MESSAGES"])) {
             const embed = new Discord.MessageEmbed()
             embed.setColor(0xFF0000)
             embed.setDescription("❌ Check My Permissions. [Missing Permission:- Manage Messages]")
-            return message.channel.send({ embeds: [embed] })
+            return send(message, {
+                embeds: [embed],
+                ephemeral: true
+            }, true)
         }
 
         if (!message.channel.permissionsFor(message.guild.me).has("MANAGE_MESSAGES")) {
             const embed = new Discord.MessageEmbed()
             embed.setColor(0xFF0000)
             embed.setDescription("❌ I don't have permission in this channel! [Missing Permission:- Manage Messages]")
-            return message.channel.send({ embeds: [embed] })
+            return send(message, {
+                embeds: [embed],
+                ephemeral: true
+            }, true)
         }
 
         var embed = new Discord.MessageEmbed()
 
-        var member;
-        var mention = args[0];
-        if (args[0]) {
-            try {
-                if (message.mentions.repliedUser) {
-                    if (mention.startsWith('<@') && mention.endsWith('>')) {
-                        mention = mention.slice(2, -1);
-
-                        if (mention.startsWith('!')) {
-                            mention = mention.slice(1);
-                        }
-                        member = await message.guild.members.fetch(mention)
-                    } else {
-                        member = message.mentions.members.get(Array.from(message.mentions.members.keys())[1]) || await message.guild.members.fetch(args[0]).catch(error => console.log()) || await bot.users.fetch(args[0]).catch(error => console.log())
-                    }
-                } else {
-                    member = message.mentions.members.first() || await message.guild.members.fetch(args[0]).catch(error => console.log()) || await bot.users.fetch(args[0]).catch(error => console.log())
-                }
-                if (!member) return message.channel.send(`<@${message.author.id}>, Invalid User!`);
-            } catch (error) {
-                if (!member) return message.channel.send(`<@${message.author.id}>, Invalid User!`);
-            }
-        } else return message.channel.send(`<@${message.author.id}>,` + "You Need To Mention A User!");
+        var member = await getMember(bot, args, options, message, author, false, false, 0, true)
 
         if (member) {
             const guildMember = message.guild.members.cache.get(member.id);
@@ -66,19 +70,18 @@ module.exports = {
                 if (member.permissions.has("ADMINISTRATOR")) {
                     embed.setColor(0xFF0000)
                     embed.setDescription("❌ You can not ban an Admin.This person seems to be an Admin of this server.")
-                    return message.channel.send({
-                        embeds: [embed]
-                    }).then(m => setTimeout(() => m.delete().catch(error => console.log(error)), 15000));
+                    return send(message, {
+                        embeds: [embed],
+                        ephemeral: true
+                    })
                 }
             }
         }
         //defining member who will get a warn and fetching id of him so member will be id of user mentioned
 
 
-        let reason2 = args
-        reason2.shift()
-        let reason = reason2.join(' ')
-
+        let reason = options[1]
+        var input = options[0]
 
         if (!reason) {
             reason = "None"
@@ -86,17 +89,18 @@ module.exports = {
 
 
         async function sendDm() {
-            const msg = "You got banned from" + message.guild.name + "\n" + "By: " + message.author.username + "\n" + "Reason: " + reason
+            const msg = "You got banned from " + message.guild.name + "\n" + "By: " + author.username + "\n" + "Reason: " + reason
             embed.setColor(0x00FFFF)
             embed.setThumbnail(bot.user.displayAvatarURL())
             embed.setDescription(msg);
-            embed.setFooter(message.author.tag, message.author.displayAvatarURL())
+            embed.setFooter({ text: author.tag, iconURL: author.displayAvatarURL() })
             embed.setTimestamp()
-            await bot.users.cache.get(member.id).send({
-                embeds: [embed]
+            var mentionedUser = bot.users.cache.get(member.id)
+            await mentionedUser.send({
+                embeds: [embed1],
             }).catch(error => {
                 console.log(error)
-            });
+            })
 
         }
 
@@ -107,9 +111,11 @@ module.exports = {
             const guildMember = message.guild.members.cache.get(member.id);
             var fetchBans = await message.guild.bans.fetch();
             var currentBan = fetchBans.get(member.id)
-            if (currentBan) return message.channel.send(`<@${message.author.id}>, This user is already banned from the server`)
+            if (currentBan) return message.channel.send(`<@${author.id}>, This user is already banned from the server`)
 
-            await message.guild.bans.create(member, { reason: reason }).then((user) => {
+            await message.guild.bans.create(member, {
+                reason: reason
+            }).then((user) => {
                 {
                     var embed1 = new Discord.MessageEmbed()
                     embed1.setColor(0x00FFFF)
@@ -119,38 +125,47 @@ module.exports = {
                         embed1.setDescription(`<:Bluecheckmark:754538270028726342> ***Successfully Banned ${user.username}#${user.discriminator}*** | **${reason}**`)
                     }
 
-                    return message.channel.send({ embeds: [embed1] })
+                    return send(message, {
+                        embeds: [embed1]
+                    })
                 }
 
             }).catch(err => {
-                message.channel.send('I was unable to ban the member');
+                send(message, { content: 'I was unable to ban the member' });
                 console.log(err);
             });
         }
 
         async function guildBan() {
-            if (isNaN(args[0])) return message.channel.send(`<@${message.author.id}>, Either the user is not present in the server or the user doesn't exist on the discord, try to use the user id instead`)
+            if (isNaN(input)) return send(message, { content: `Either the user is not present in the server or the user doesn't exist on the discord, try to use the user id instead` }
+                , false)
             var fetchBans = await message.guild.bans.fetch();
-            // console.log(args[0])
-            var currentBan = fetchBans.get(args[0])
+            // console.log(input)
+            var currentBan = fetchBans.get(input)
 
-            if (currentBan) return message.channel.send(`<@${message.author.id}>, This user is already banned from the server`)
+            if (currentBan) return send(message, { content: `This user is already banned from the server` }
+                , false)
 
-            await message.guild.bans.create(args[0]).then((user) => {
+            await message.guild.bans.create(input).then((user) => {
                 {
                     var embed1 = new Discord.MessageEmbed()
                     embed1.setColor(0x00FFFF)
-                    embed1.description(`<:Bluecheckmark:754538270028726342> ***Successfully Banned ${user}*** | **${reason}**`)
-                    return message.channel.send({ embeds: [embed1] })
+                    embed1.setDescription(`<:Bluecheckmark:754538270028726342> ***Successfully Banned ${user.username}#${user.discriminator}*** | **${reason}**`)
+                    return send(message, {
+                        embeds: [embed1]
+                    }, false)
                 }
 
             }).catch(err => {
-                message.channel.send('I was unable to ban the member');
+                send(message, { content: 'I was unable to ban the member' }
+                    , false);
                 console.log(err);
             });
         }
 
-        await message.delete().catch(error => console.log(error))
+        if (message.type == "DEFAULT" || message.type == "REPLY") {
+            await message.delete().catch(error => console.log(error))
+        }
 
         if (member) {
             const guildMember = message.guild.members.cache.get(member.id);
