+++ commands/Moderation/mute.js
@@ -3,6 +3,8 @@ const Discord = require("discord.js");
 const ms = require('ms');
 const moment = require('moment')
 const server = require("../../model/server.js")
+const send = require("../../utils/sendMessage.js")
+const getMember = require("../../utils/getMember.js");
 
 module.exports = {
     name: "mute",
@@ -14,7 +16,27 @@ module.exports = {
     example: "=mute @Real Warrior perm Abuse , =mute @Yashu 10m Spamming , =mute @Shander 1d Emoji Spamming",
     permission: ["MANAGE_MESSAGES"],
     botreq: "Embed Links, Manage Roles, Manage Messages",
-    run: async (bot, message, args) => {
+    options: [{
+        name: "user",
+        description: "User to be muted",
+        required: true,
+        type: 6, //https://discord.com/developers/docs/interactions/application-commands#application-command-object-application-command-option-structure
+        req: "string"
+    }, {
+        name: "time",
+        description: "The duration for which the user will be muted, use perm for permanant mute",
+        required: true,
+        type: 3, //https://discord.com/developers/docs/interactions/application-commands#application-command-object-application-command-option-structure
+        req: "string"
+    }, {
+        name: "reason",
+        description: "Reason why the member is being muted",
+        required: false,
+        type: 3, //https://discord.com/developers/docs/interactions/application-commands#application-command-object-application-command-option-structure
+        req: "reason"
+    }],
+    run: async (bot, message, args, options, author) => {
+
 
 
         //perission code to run when it doesnt matter even if errormsg system is on/off-
@@ -23,77 +45,72 @@ module.exports = {
             const embed = new Discord.MessageEmbed()
             embed.setColor(0xFF0000)
             embed.setDescription("❌ Check My Permissions. [Missing Permission:- Manage Roles]")
-            return message.channel.send({ embeds: [embed] })
+            return send(message, {
+                embeds: [embed],
+                ephemeral: true
+            }, false)
         }
 
         if (!message.guild.me.permissions.has(["MANAGE_MESSAGES"])) {
             const embed = new Discord.MessageEmbed()
             embed.setColor(0xFF0000)
             embed.setDescription("❌ Check My Permissions. [Missing Permission:- Manage Messages]")
-            return message.channel.send({ embeds: [embed] })
+            return send(message, {
+                embeds: [embed],
+                ephemeral: true
+            }, false)
         }
 
         if (!message.channel.permissionsFor(message.guild.me).has("MANAGE_MESSAGES")) {
             const embed = new Discord.MessageEmbed()
             embed.setColor(0xFF0000)
             embed.setDescription("❌ I don't have permission in this channel! [Missing Permission:- Manage Messages]")
-            return message.channel.send({ embeds: [embed] })
+            return send(message, {
+                embeds: [embed],
+                ephemeral: true
+            }, false)
         }
-
         //main code-
 
 
-        var person;
-        var mention = args[0];
-        if (args[0]) {
-            try {
-                if (message.mentions.repliedUser) {
-                    if (mention.startsWith('<@') && mention.endsWith('>')) {
-                        mention = mention.slice(2, -1);
+        var person = await getMember(bot, args, options, message, author, false, false, 0, false)
 
-                        if (mention.startsWith('!')) {
-                            mention = mention.slice(1);
-                        }
-                        person = await message.guild.members.fetch(mention)
-                    } else {
-                        person = message.mentions.members.get(Array.from(message.mentions.members.keys())[1]) || await message.guild.members.fetch(args[0]).catch(error => console.log())
-                    }
-                } else {
-                    person = message.mentions.members.first() || await message.guild.members.fetch(args[0]).catch(error => console.log())
-                }
-                if (!person) return message.channel.send(`<@${message.author.id}>, Invalid User!`);
-            } catch (error) {
-                if (!person) return message.channel.send(`<@${message.author.id}>, Invalid User!`);
-            }
-        } else {
-            if (!person) return message.channel.send(`<@${message.author.id}> , You Need To Mention A User!`);
+        if (!person) {
+            return send(message, { content: `You Need To Mention A User!` }, true)
         }
 
-        if (person.permissions.has("ADMINISTRATOR")) return message.reply("❌ You can not mute an Admin. This person seems to be an Admin of this server.")
+        if (person.permissions.has("ADMINISTRATOR")) return send(message, { content: "❌ You can not mute an Admin. This person seems to be an Admin of this server." }, true)
         const Createdd = Date.now()
 
-        let time = args[1];
-        let reason = args.slice(2).join(" ")
+        let time = options[1];
+        let reason = options[2]
 
-        if (!time) return message.reply("You need to tell time");
+        if (time.toLowerCase() == "perm" || time.toLowerCase() == "permanent") {
+            time = "perm"
+        }
 
         time = ms(time)
-        if (time < 60000) return message.channel.send("Minimum time limit is 1 minute!")
 
         if (!time) {
             time = "perm"
-            reason = args.slice(1).join(" ")
+            if (message.type == "DEFAULT" || message.type == "REPLY") {
+                reason = options[1]
+            }
         }
 
-        if (!reason) return message.reply("You didn't specify a reason!");
 
+        if (!isNaN(time)) {
+            if (time < 60000) return send(message, { content: "Minimum time limit is 1 minute!" }, true)
+        }
+
+        if (!reason) reason = "None"
 
         // console.log(time)
         var muterole = message.guild.roles.cache.find(role => role.name === 'Muted');
 
         if (!muterole) {
             try {
-                message.channel.send(`Muted role is not found, Attemting to create a Muted role.`)
+                send(message, { content: `Muted role is not found, Attemting to create a Muted role.` }, false)
 
                 let rolecreation = await message.guild.roles.create({
                     name: 'Muted',
@@ -115,7 +132,7 @@ module.exports = {
                         SPEAK: false
                     })
                 });
-                message.channel.send("The muted role has been successfully created")
+                send(message, { content: "The muted role has been successfully created" }, false)
 
             } catch (error) {
                 console.log(error)
@@ -130,14 +147,15 @@ module.exports = {
             const embed = new Discord.MessageEmbed()
             embed.setDescription("Please Check My Permission, Maybe my role isn't higher enough in order to give a role to the user!")
             embed.setColor(0xff4a1f)
-            return await message.channel.send({ embeds: [embed] })
+            return send(message, {
+                embeds: [embed],
+                ephemeral: true
+            }, false)
         }
 
-        // console.log(reason)
 
         // if (!message.guild.me.hasPermission("MANAGE_ROLES")) return message.channel.send(embed2)
-
-        if (person.roles.cache.some(r => r.name === "Muted")) return message.channel.send(`<@${message.author.id}>, ${person.user.tag} is already muted`)
+        if (person.roles.cache.some(r => r.name === "Muted")) return send(message, { content: `${person.user.tag} is already muted` }, true)
         // console.log(person.roles.cache.some(r => r.name === "Muted"))
 
 
@@ -152,7 +170,7 @@ module.exports = {
             }, async (err, user) => {
 
                 var muteObj = {
-                    administrator: message.author.tag,
+                    administrator: author.tag,
                     reason: reason,
                     duration: time,
                     type: "Mute",
@@ -184,28 +202,37 @@ module.exports = {
                     const embed3 = new Discord.MessageEmbed()
                     embed3.setColor(0x00FFFF)
                     embed3.setDescription(`<:Bluecheckmark:754538270028726342> ***${person.user.tag} has been muted permanantely*** | **${reason}** `);
-                    message.channel.send({ embeds: [embed3] });
+                    send(message, {
+                        embeds: [embed3]
+                    }, false);
                     const embed69 = new Discord.MessageEmbed()
                     embed69.setColor(0x00FFFF)
                     embed69.setDescription(`<:Bluecheckmark:754538270028726342> ***You have been muted permanantely in ${message.guild.name}*** | **${reason}**`);
-                    person.send({ embeds: [embed69] }).catch(error => console.log(error))
+                    try {
+                        bot.users.cache.get(person.id).send({
+                            embeds: [embed69]
+                        });
+                    } catch (error) {
+                        console.log(error)
+                    }
                 } else {
                     person.roles.add(message.guild.roles.cache.find(role => role.name === 'Muted'))
 
                     const embed99 = new Discord.MessageEmbed()
                     embed99.setColor(0x00FFFF)
                     embed99.setDescription(`<:Bluecheckmark:754538270028726342> ***${person.user.tag} has been muted for ${ms(time)}*** | **${reason}** `);
-                    message.channel.send({ embeds: [embed99] });
+                    send(message, {
+                        embeds: [embed99]
+                    }, false);
                     const embed420 = new Discord.MessageEmbed()
                     embed420.setColor(0x00FFFF)
                     embed420.setDescription(`<:Bluecheckmark:754538270028726342> ***You have been muted for ${ms(time)} in ${message.guild.name}*** | **${reason}**`);
-
-                    await bot.users.cache.get(person.id).send({
-                        embeds: [embed420]
+                    var mentionedUser = bot.users.cache.get(person.id)
+                    await mentionedUser.send({
+                        embeds: [embed420],
                     }).catch(error => {
                         console.log(error)
-                    });
-
+                    })
                     // console.log(time)
                     setTimeout(async () => {
                         var guildMember = message.guild.members.fetch(person.id).catch(error => console.log(error))
@@ -257,7 +284,8 @@ module.exports = {
             })
         }
         doMute()
-
-        await message.delete().catch(error => console.log(error))
+        if (message.type == "DEFAULT" || message.type == "REPLY") {
+            await message.delete().catch(error => console.log(error))
+        }
     }
 }
